Revert "Created shared config for Travis-CI builds (#18)"
This reverts commit 0d6c0c59a80e34a20288bd580ac0242969916ff3.
