+++ vsts/config/utils/browsers.js
